Added a Dirac algebra class using boost's ublas.
Since we want fast matrix operations for the fermions, I thought it
would be a good idea to use blas of some sort to implement them. Of
course, the Dirac algebra is a good point to start. Note you have to
pass the --with-boost-srcdir=path/to/boost command line option to the
configure script now or install boost for things to work now. You can
get boost at www.boost.org.