Cleaned up some of the mess in kernels.
The idea is to take ONE class that extracts all important info from
the Field class and this is then used in all the other Kernels. In
such a way one only writes Kernel<Field> instead of Kernel<Field, DIM,
ORD> or some nonsense like that. The goal is however to find a better
way yet with some sort of inheritance or so.